don't resort processed centroids
The processed centroids are already sorted. Instead of resorting them
together with the unprocessed centroids, we sort just the unprocessed
centroids and do an "inline merge" of the two sorted lists.

This method will also be useful when merging digests.

```
name            old time/op  new time/op  delta
TDigest_Add-10  80.0ms ± 1%  69.0ms ± 1%  -13.74%  (p=0.008 n=5+5)
```

+++ tdigest_test.go
@@ -609,7 +609,7 @@ func TestAccuracy(t *testing.T) {
 	if testing.Short() {
 		t.Skip("skipping test in short mode")
 	}
-	for _, delta := range []float64{100, 1000} {
+	for _, delta := range []int{100, 1000} {
 		t.Run(fmt.Sprintf("delta=%v", delta), func(t *testing.T) {
 			for valDistName, valDistFn := range testDistributions {
 				for valModName, valModFn := range testDistributionModifiers {
